+++ Assets/Scripts/Classes/DoobiesScripts/Instances/VangurrInstance.cs
@@ -8,7 +8,7 @@ public class VangurrInstance : CombatantInstance
     public override string CharacterName => _so.vangurrName;
     public override int CurrentHealth { get; set; }
 
-    public override int CurrentDefence { get; set; }
+    public override float CurrentDefence { get; set; }
 
     public override List<SkillSO> GetAllSkills() => new List<SkillSO>(_so.baseSkills);
